Fix compilation warnings when targeting x64

+++ FlexEngine/include/Graphics/Vulkan/VulkanHelpers.hpp
@@ -499,7 +499,7 @@ namespace flex
 
 			u32 Count()
 			{
-				return values.size();
+				return (u32)values.size();
 			}
 
 			iter begin()
